simplitting out the formats to be handled separately.
